Handle the warning for trezor-crypto.

+++ trezor-crypto/src/bip39.c
@@ -58,9 +58,9 @@ bool mnemonic_from_data(const uint8_t *data, size_t len, char* mnemonic) {
 	// data
 	memcpy(bits, data, len);
 
-	int mlen = len * 3 / 4;
+	size_t mlen = len * 3 / 4;
 
-	int i, j, idx;
+	size_t i, j, idx;
 	char *p = mnemonic;
 	for (i = 0; i < mlen; i++) {
 		idx = 0;
@@ -203,8 +203,8 @@ char *normalize_mnemonic(const char *mnemonic) {
 void mnemonic_to_seed(const char *mnemonic, const char *passphrase, uint8_t seed[512 / 8], void (*progress_callback)(uint32_t current, uint32_t total))
 {
 	char *normalized = normalize_mnemonic(mnemonic);
-	int normalizedlen = strlen(normalized);
-	int passphraselen = strnlen(passphrase, 256);
+	size_t normalizedlen = strlen(normalized);
+	size_t passphraselen = strnlen(passphrase, 256);
 	uint8_t salt[8 + 256];
 	memcpy(salt, "mnemonic", 8);
 	memcpy(salt + 8, passphrase, passphraselen);
